+++ pytissueoptics/__init__.py
@@ -1,4 +1,2 @@
 from .scene import *
 from .rayscattering import *
-from .rayscattering import tissues
-from .rayscattering.tissues import ScatteringScene
