+++ pytissueoptics/rayscattering/__init__.py
@@ -2,13 +2,14 @@ from .photon import Photon
 from .materials import ScatteringMaterial
 from .energyLogging import EnergyLogger
 from .source import PencilPointSource, IsotropicPointSource, DirectionalSource, DivergentSource
+from .scatteringScene import ScatteringScene
 from .display.viewer import Viewer, PointCloudStyle, Visibility, ViewGroup, Direction
 from .display.views import View2DProjection, View2DProjectionX, View2DProjectionY, View2DProjectionZ, \
     View2DSurface, View2DSurfaceX, View2DSurfaceY, View2DSurfaceZ, View2DSlice, View2DSliceX, View2DSliceY, View2DSliceZ
 
 
 __all__ = ["Photon", "ScatteringMaterial", "PencilPointSource", "IsotropicPointSource", "DirectionalSource",
-           "DivergentSource", "EnergyLogger", "Viewer", "PointCloudStyle", "Visibility", "ViewGroup",
+           "DivergentSource", "EnergyLogger", "ScatteringScene", "Viewer", "PointCloudStyle", "Visibility", "ViewGroup",
            "Direction", "View2DProjection", "View2DProjectionX", "View2DProjectionY", "View2DProjectionZ",
            "View2DSurface", "View2DSurfaceX", "View2DSurfaceY", "View2DSurfaceZ", "View2DSlice", "View2DSliceX",
-           "View2DSliceY", "View2DSliceZ"]
+           "View2DSliceY", "View2DSliceZ", "samples"]
