+++ pytissueoptics/rayscattering/samples/__init__.py
@@ -1,3 +1,2 @@
 from .phantomTissue import PhantomTissue
 from .infiniteTissue import InfiniteTissue
-from .scatteringScene import ScatteringScene
