+++ pytissueoptics/rayscattering/tests/tissues/testScatteringScene.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 
 from mockito import mock, verify, when
 
-from pytissueoptics.rayscattering.tissues import ScatteringScene
+from pytissueoptics.rayscattering.samples import ScatteringScene
 from pytissueoptics.rayscattering.materials import ScatteringMaterial
 from pytissueoptics.scene.solids import Cuboid
 from pytissueoptics.scene.viewer import MayaviViewer
